start implementing scene stuff
